+++ extensions/git/package.json
@@ -20,8 +20,7 @@
     "scmValidation",
     "tabInputTextMerge",
     "timeline",
-    "contribMergeEditorMenus",
-    "extensionLog"
+    "contribMergeEditorMenus"
   ],
   "categories": [
     "Other"
