+++ extensions/github-authentication/package.json
@@ -52,8 +52,7 @@
     }
   },
   "enabledApiProposals": [
-    "telemetryLogger",
-    "extensionLog"
+    "telemetryLogger"
   ],
   "aiKey": "0c6ae279ed8443289764825290e4f9e2-1a736e7c-1324-4338-be46-fc2a58ae4d14-7255",
   "main": "./out/extension.js",
