+++ extensions/github-authentication/tsconfig.json
@@ -12,7 +12,6 @@
 	},
 	"include": [
 		"src/**/*",
-		"../../src/vscode-dts/vscode.d.ts",
-		"../../src/vscode-dts/vscode.proposed.extensionLog.d.ts"
+		"../../src/vscode-dts/vscode.d.ts"
 	]
 }
