+++ extensions/microsoft-authentication/package.json
@@ -17,8 +17,7 @@
   ],
   "enabledApiProposals": [
     "telemetryLogger",
-    "idToken",
-    "extensionLog"
+    "idToken"
   ],
   "capabilities": {
     "virtualWorkspaces": true,
