+++ extensions/microsoft-authentication/tsconfig.json
@@ -22,7 +22,6 @@
 	"include": [
 		"src/**/*",
 		"../../src/vscode-dts/vscode.d.ts",
-		"../../src/vscode-dts/vscode.proposed.idToken.d.ts",
-		"../../src/vscode-dts/vscode.proposed.extensionLog.d.ts"
+		"../../src/vscode-dts/vscode.proposed.idToken.d.ts"
 	]
 }
