+++ src/vs/workbench/api/common/extHostOutput.ts
@@ -20,7 +20,6 @@ import { isString } from 'vs/base/common/types';
 import { FileSystemProviderErrorCode, toFileSystemProviderErrorCode } from 'vs/platform/files/common/files';
 import { Emitter } from 'vs/base/common/event';
 import { DisposableStore } from 'vs/base/common/lifecycle';
-import { checkProposedApiEnabled } from 'vs/workbench/services/extensions/common/extensions';
 
 class ExtHostOutputChannel extends AbstractMessageLogger implements vscode.LogOutputChannel {
 
@@ -144,9 +143,6 @@ export class ExtHostOutputService implements ExtHostOutputServiceShape {
 			throw new Error('illegal argument `name`. must not be falsy');
 		}
 		const log = typeof options === 'object' && options.log;
-		if (log) {
-			checkProposedApiEnabled(extension, 'extensionLog');
-		}
 		const languageId = isString(options) ? options : undefined;
 		if (isString(languageId) && !languageId.trim()) {
 			throw new Error('illegal argument `languageId`. must not be empty');
